fix: The bullet will occasionally be bounced to the wrong position
The reason is probably because the initial position of the bullet overlaps with the position of the launcher, and there is a collision between the two, which causes the position of the bullet to fly. Canceling the collision of objects marked Harm and Harmful and relying only on raycast for collision detection seems to solve this problem.